Add external CSS styling to <div> element
This commit adds external CSS styling to the <div> element. An external CSS file named "styles.css" has been added to the repo and a link to this file has been created in the <header> section of index.html.

The style changes made to the <div> type elements are listed below.
- White text
- Red background
- Center aligned text
- Size 32px font